split onroad
