+++ selfdrive/ui/qt/onroad/annotated_camera.cc
@@ -1,310 +1,13 @@
-#include "selfdrive/ui/qt/onroad.h"
 
+#include "selfdrive/ui/qt/onroad/annotated_camera.h"
+
+#include <QPainter>
 #include <algorithm>
 #include <cmath>
-#include <map>
-#include <memory>
-#include <sstream>
-
-#include <QDebug>
-#include <QMouseEvent>
 
 #include "common/swaglog.h"
-#include "common/timing.h"
+#include "selfdrive/ui/qt/onroad/buttons.h"
 #include "selfdrive/ui/qt/util.h"
-#ifdef ENABLE_MAPS
-#include "selfdrive/ui/qt/maps/map_helpers.h"
-#include "selfdrive/ui/qt/maps/map_panel.h"
-#endif
-
-static void drawIcon(QPainter &p, const QPoint &center, const QPixmap &img, const QBrush &bg, float opacity) {
-  p.setRenderHint(QPainter::Antialiasing);
-  p.setOpacity(1.0);  // bg dictates opacity of ellipse
-  p.setPen(Qt::NoPen);
-  p.setBrush(bg);
-  p.drawEllipse(center, btn_size / 2, btn_size / 2);
-  p.setOpacity(opacity);
-  p.drawPixmap(center - QPoint(img.width() / 2, img.height() / 2), img);
-  p.setOpacity(1.0);
-}
-
-OnroadWindow::OnroadWindow(QWidget *parent) : QWidget(parent) {
-  QVBoxLayout *main_layout  = new QVBoxLayout(this);
-  main_layout->setMargin(UI_BORDER_SIZE);
-  QStackedLayout *stacked_layout = new QStackedLayout;
-  stacked_layout->setStackingMode(QStackedLayout::StackAll);
-  main_layout->addLayout(stacked_layout);
-
-  nvg = new AnnotatedCameraWidget(VISION_STREAM_ROAD, this);
-
-  QWidget * split_wrapper = new QWidget;
-  split = new QHBoxLayout(split_wrapper);
-  split->setContentsMargins(0, 0, 0, 0);
-  split->setSpacing(0);
-  split->addWidget(nvg);
-
-  if (getenv("DUAL_CAMERA_VIEW")) {
-    CameraWidget *arCam = new CameraWidget("camerad", VISION_STREAM_ROAD, true, this);
-    split->insertWidget(0, arCam);
-  }
-
-  if (getenv("MAP_RENDER_VIEW")) {
-    CameraWidget *map_render = new CameraWidget("navd", VISION_STREAM_MAP, false, this);
-    split->insertWidget(0, map_render);
-  }
-
-  stacked_layout->addWidget(split_wrapper);
-
-  alerts = new OnroadAlerts(this);
-  alerts->setAttribute(Qt::WA_TransparentForMouseEvents, true);
-  stacked_layout->addWidget(alerts);
-
-  // setup stacking order
-  alerts->raise();
-
-  setAttribute(Qt::WA_OpaquePaintEvent);
-  QObject::connect(uiState(), &UIState::uiUpdate, this, &OnroadWindow::updateState);
-  QObject::connect(uiState(), &UIState::offroadTransition, this, &OnroadWindow::offroadTransition);
-  QObject::connect(uiState(), &UIState::primeChanged, this, &OnroadWindow::primeChanged);
-}
-
-void OnroadWindow::updateState(const UIState &s) {
-  if (!s.scene.started) {
-    return;
-  }
-
-  if (s.scene.map_on_left) {
-    split->setDirection(QBoxLayout::LeftToRight);
-  } else {
-    split->setDirection(QBoxLayout::RightToLeft);
-  }
-
-  alerts->updateState(s);
-  nvg->updateState(s);
-
-  QColor bgColor = bg_colors[s.status];
-  if (bg != bgColor) {
-    // repaint border
-    bg = bgColor;
-    update();
-  }
-}
-
-void OnroadWindow::mousePressEvent(QMouseEvent* e) {
-#ifdef ENABLE_MAPS
-  if (map != nullptr) {
-    bool sidebarVisible = geometry().x() > 0;
-    bool show_map = !sidebarVisible;
-    map->setVisible(show_map && !map->isVisible());
-  }
-#endif
-  // propagation event to parent(HomeWindow)
-  QWidget::mousePressEvent(e);
-}
-
-void OnroadWindow::createMapWidget() {
-#ifdef ENABLE_MAPS
-  auto m = new MapPanel(get_mapbox_settings());
-  map = m;
-  QObject::connect(m, &MapPanel::mapPanelRequested, this, &OnroadWindow::mapPanelRequested);
-  QObject::connect(nvg->map_settings_btn, &MapSettingsButton::clicked, m, &MapPanel::toggleMapSettings);
-  nvg->map_settings_btn->setEnabled(true);
-
-  m->setFixedWidth(topWidget(this)->width() / 2 - UI_BORDER_SIZE);
-  split->insertWidget(0, m);
-  // hidden by default, made visible when navRoute is published
-  m->setVisible(false);
-#endif
-}
-
-void OnroadWindow::offroadTransition(bool offroad) {
-#ifdef ENABLE_MAPS
-  if (!offroad) {
-    if (map == nullptr && (uiState()->hasPrime() || !MAPBOX_TOKEN.isEmpty())) {
-      createMapWidget();
-    }
-  }
-#endif
-  alerts->clear();
-}
-
-void OnroadWindow::primeChanged(bool prime) {
-#ifdef ENABLE_MAPS
-  if (map && (!prime && MAPBOX_TOKEN.isEmpty())) {
-    nvg->map_settings_btn->setEnabled(false);
-    nvg->map_settings_btn->setVisible(false);
-    map->deleteLater();
-    map = nullptr;
-  } else if (!map && (prime || !MAPBOX_TOKEN.isEmpty())) {
-    createMapWidget();
-  }
-#endif
-}
-
-void OnroadWindow::paintEvent(QPaintEvent *event) {
-  QPainter p(this);
-  p.fillRect(rect(), QColor(bg.red(), bg.green(), bg.blue(), 255));
-}
-
-// ***** onroad widgets *****
-
-// OnroadAlerts
-
-void OnroadAlerts::updateState(const UIState &s) {
-  Alert a = getAlert(*(s.sm), s.scene.started_frame);
-  if (!alert.equal(a)) {
-    alert = a;
-    update();
-  }
-}
-
-void OnroadAlerts::clear() {
-  alert = {};
-  update();
-}
-
-OnroadAlerts::Alert OnroadAlerts::getAlert(const SubMaster &sm, uint64_t started_frame) {
-  const cereal::ControlsState::Reader &cs = sm["controlsState"].getControlsState();
-  const uint64_t controls_frame = sm.rcv_frame("controlsState");
-
-  Alert a = {};
-  if (controls_frame >= started_frame) {  // Don't get old alert.
-    a = {cs.getAlertText1().cStr(), cs.getAlertText2().cStr(),
-         cs.getAlertType().cStr(), cs.getAlertSize(), cs.getAlertStatus()};
-  }
-
-  if (!sm.updated("controlsState") && (sm.frame - started_frame) > 5 * UI_FREQ) {
-    const int CONTROLS_TIMEOUT = 5;
-    const int controls_missing = (nanos_since_boot() - sm.rcv_time("controlsState")) / 1e9;
-
-    // Handle controls timeout
-    if (controls_frame < started_frame) {
-      // car is started, but controlsState hasn't been seen at all
-      a = {tr("openpilot Unavailable"), tr("Waiting for controls to start"),
-           "controlsWaiting", cereal::ControlsState::AlertSize::MID,
-           cereal::ControlsState::AlertStatus::NORMAL};
-    } else if (controls_missing > CONTROLS_TIMEOUT && !Hardware::PC()) {
-      // car is started, but controls is lagging or died
-      if (cs.getEnabled() && (controls_missing - CONTROLS_TIMEOUT) < 10) {
-        a = {tr("TAKE CONTROL IMMEDIATELY"), tr("Controls Unresponsive"),
-             "controlsUnresponsive", cereal::ControlsState::AlertSize::FULL,
-             cereal::ControlsState::AlertStatus::CRITICAL};
-      } else {
-        a = {tr("Controls Unresponsive"), tr("Reboot Device"),
-             "controlsUnresponsivePermanent", cereal::ControlsState::AlertSize::MID,
-             cereal::ControlsState::AlertStatus::NORMAL};
-      }
-    }
-  }
-  return a;
-}
-
-void OnroadAlerts::paintEvent(QPaintEvent *event) {
-  if (alert.size == cereal::ControlsState::AlertSize::NONE) {
-    return;
-  }
-  static std::map<cereal::ControlsState::AlertSize, const int> alert_heights = {
-    {cereal::ControlsState::AlertSize::SMALL, 271},
-    {cereal::ControlsState::AlertSize::MID, 420},
-    {cereal::ControlsState::AlertSize::FULL, height()},
-  };
-  int h = alert_heights[alert.size];
-
-  int margin = 40;
-  int radius = 30;
-  if (alert.size == cereal::ControlsState::AlertSize::FULL) {
-    margin = 0;
-    radius = 0;
-  }
-  QRect r = QRect(0 + margin, height() - h + margin, width() - margin*2, h - margin*2);
-
-  QPainter p(this);
-
-  // draw background + gradient
-  p.setPen(Qt::NoPen);
-  p.setCompositionMode(QPainter::CompositionMode_SourceOver);
-  p.setBrush(QBrush(alert_colors[alert.status]));
-  p.drawRoundedRect(r, radius, radius);
-
-  QLinearGradient g(0, r.y(), 0, r.bottom());
-  g.setColorAt(0, QColor::fromRgbF(0, 0, 0, 0.05));
-  g.setColorAt(1, QColor::fromRgbF(0, 0, 0, 0.35));
-
-  p.setCompositionMode(QPainter::CompositionMode_DestinationOver);
-  p.setBrush(QBrush(g));
-  p.drawRoundedRect(r, radius, radius);
-  p.setCompositionMode(QPainter::CompositionMode_SourceOver);
-
-  // text
-  const QPoint c = r.center();
-  p.setPen(QColor(0xff, 0xff, 0xff));
-  p.setRenderHint(QPainter::TextAntialiasing);
-  if (alert.size == cereal::ControlsState::AlertSize::SMALL) {
-    p.setFont(InterFont(74, QFont::DemiBold));
-    p.drawText(r, Qt::AlignCenter, alert.text1);
-  } else if (alert.size == cereal::ControlsState::AlertSize::MID) {
-    p.setFont(InterFont(88, QFont::Bold));
-    p.drawText(QRect(0, c.y() - 125, width(), 150), Qt::AlignHCenter | Qt::AlignTop, alert.text1);
-    p.setFont(InterFont(66));
-    p.drawText(QRect(0, c.y() + 21, width(), 90), Qt::AlignHCenter, alert.text2);
-  } else if (alert.size == cereal::ControlsState::AlertSize::FULL) {
-    bool l = alert.text1.length() > 15;
-    p.setFont(InterFont(l ? 132 : 177, QFont::Bold));
-    p.drawText(QRect(0, r.y() + (l ? 240 : 270), width(), 600), Qt::AlignHCenter | Qt::TextWordWrap, alert.text1);
-    p.setFont(InterFont(88));
-    p.drawText(QRect(0, r.height() - (l ? 361 : 420), width(), 300), Qt::AlignHCenter | Qt::TextWordWrap, alert.text2);
-  }
-}
-
-// ExperimentalButton
-ExperimentalButton::ExperimentalButton(QWidget *parent) : experimental_mode(false), engageable(false), QPushButton(parent) {
-  setFixedSize(btn_size, btn_size);
-
-  engage_img = loadPixmap("../assets/img_chffr_wheel.png", {img_size, img_size});
-  experimental_img = loadPixmap("../assets/img_experimental.svg", {img_size, img_size});
-  QObject::connect(this, &QPushButton::clicked, this, &ExperimentalButton::changeMode);
-}
-
-void ExperimentalButton::changeMode() {
-  const auto cp = (*uiState()->sm)["carParams"].getCarParams();
-  bool can_change = hasLongitudinalControl(cp) && params.getBool("ExperimentalModeConfirmed");
-  if (can_change) {
-    params.putBool("ExperimentalMode", !experimental_mode);
-  }
-}
-
-void ExperimentalButton::updateState(const UIState &s) {
-  const auto cs = (*s.sm)["controlsState"].getControlsState();
-  bool eng = cs.getEngageable() || cs.getEnabled();
-  if ((cs.getExperimentalMode() != experimental_mode) || (eng != engageable)) {
-    engageable = eng;
-    experimental_mode = cs.getExperimentalMode();
-    update();
-  }
-}
-
-void ExperimentalButton::paintEvent(QPaintEvent *event) {
-  QPainter p(this);
-  QPixmap img = experimental_mode ? experimental_img : engage_img;
-  drawIcon(p, QPoint(btn_size / 2, btn_size / 2), img, QColor(0, 0, 0, 166), (isDown() || !engageable) ? 0.6 : 1.0);
-}
-
-
-// MapSettingsButton
-MapSettingsButton::MapSettingsButton(QWidget *parent) : QPushButton(parent) {
-  setFixedSize(btn_size, btn_size);
-  settings_img = loadPixmap("../assets/navigation/icon_directions_outlined.svg", {img_size, img_size});
-
-  // hidden by default, made visible if map is created (has prime or mapbox token)
-  setVisible(false);
-  setEnabled(false);
-}
-
-void MapSettingsButton::paintEvent(QPaintEvent *event) {
-  QPainter p(this);
-  drawIcon(p, QPoint(btn_size / 2, btn_size / 2), settings_img, QColor(0, 0, 0, 166), isDown() ? 0.6 : 1.0);
-}
-
 
 // Window that shows camera view and variety of info drawn on top
 AnnotatedCameraWidget::AnnotatedCameraWidget(VisionStreamType type, QWidget* parent) : fps_filter(UI_FREQ, 3, 1. / UI_FREQ), CameraWidget("camerad", type, true, parent) {
