+++ selfdrive/ui/qt/onroad/onroad_home.h
@@ -1,0 +1,31 @@
+#pragma once
+
+#include "selfdrive/ui/qt/onroad/alerts.h"
+#include "selfdrive/ui/qt/onroad/annotated_camera.h"
+
+class OnroadWindow : public QWidget {
+  Q_OBJECT
+
+public:
+  OnroadWindow(QWidget* parent = 0);
+  bool isMapVisible() const { return map && map->isVisible(); }
+  void showMapPanel(bool show) { if (map) map->setVisible(show); }
+
+signals:
+  void mapPanelRequested();
+
+private:
+  void createMapWidget();
+  void paintEvent(QPaintEvent *event);
+  void mousePressEvent(QMouseEvent* e) override;
+  OnroadAlerts *alerts;
+  AnnotatedCameraWidget *nvg;
+  QColor bg = bg_colors[STATUS_DISENGAGED];
+  QWidget *map = nullptr;
+  QHBoxLayout* split;
+
+private slots:
+  void offroadTransition(bool offroad);
+  void primeChanged(bool prime);
+  void updateState(const UIState &s);
+};
